+++ scripts/cli/src/commands/fully-managed-connector/show-config.sh
@@ -1,6 +1,7 @@
 connector="${args[--connector]}"
 force_rest_endpoint="${args[--force-rest-endpoint]}"
 verbose="${args[--verbose]}"
+no_clipboard="${args[--no-clipboard]}"
 
 get_ccloud_connect
 
@@ -41,7 +42,7 @@ do
                 playground config set clipboard true
             fi
 
-            if [ "$clipboard" == "true" ] || [ "$clipboard" == "" ]
+            if ( [ "$clipboard" == "true" ] || [ "$clipboard" == "" ] ) && [[ ! -n "$no_clipboard" ]]
             then
                 tmp_dir=$(mktemp -d -t ci-XXXXXXXXXX)
                 trap 'rm -rf $tmp_dir' EXIT
@@ -90,7 +91,7 @@ do
                         playground config set clipboard true
                     fi
 
-                    if [ "$clipboard" == "true" ] || [ "$clipboard" == "" ]
+                    if ( [ "$clipboard" == "true" ] || [ "$clipboard" == "" ] ) && [[ ! -n "$no_clipboard" ]]
                     then
                         tmp_dir=$(mktemp -d -t ci-XXXXXXXXXX)
                         trap 'rm -rf $tmp_dir' EXIT
